update: Treasury.sol update: Boardroom.sol update: Oracle.sol

+++ contracts/interfaces/IOracle.sol
@@ -7,5 +7,5 @@ interface IOracle {
 
     function consult(address _token, uint256 _amountIn) external view returns (uint144 amountOut);
 
-    function twap(address _token, uint256 _amountIn) external view returns (uint144 _amountOut);
+    function twap(address _token, uint256 _amountIn) external view returns (uint256 _amountOut);
 }
